@@ -1,10 +1,8 @@
 export { compose, pipe } from 'ramda';
 
 export * from './redux/actions';
-export { JASON_API_REQUEST } from './redux/actionTypes';
-export { default as middleware, middlewareFactory } from './redux/middleware';
+export * from './redux/hooks';
+export * from './redux/middleware';
 export * from './redux/selectors';
-export { default as reducer } from './redux/reducer';
-
-export { FlexiblePayload, JasonApiState } from './common-types';
-export * from './higher-order-components';
+export * from './redux/reducer';
+export * from './types';
